mid-project-347359563/services/payment-service/requirements.txt
fastapi
uvicorn
pika
pymysql
sqlalchemy
python-dotenv
pydantic
cryptography